pyproject.toml
[tool.poetry]
name = "convertable"
version = "0.2.0"
description = "CLI tools for converting between different formats such as JSON, YAML, and TOML"
authors = ["Mathew Moon <me@mathewmoon.net>"]
homepage = "https://github.com/mathewmoon/convertable"
repository = "https://github.com/mathewmoon/convertable"
readme = "README.md"
packages = [{include = "convertable", from = "."}]

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1.34.56"
pygments = "^2.17.0"
PyYAML = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.3"

[tool.poetry.scripts]
yamltojson = "convertable.cli:yamltojson"
tomltojson = "convertable.cli:tomltojson"
tomltoyaml = "convertable.cli:tomltoyaml"
convertable = "convertable.cli:convertable"

[tool.pytest.ini_options]
xfail_strict = true

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"
